fix: rebase auto login logic into App Container
자동 로그인 로직이 App 이 마운트 될때 동작하는 게 맞다고 판단하여 기존 인덱스 페이지에서 옮겼습니다
