+++ src/App.js
@@ -1,10 +1,15 @@
 import React, { useEffect } from "react";
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+import { Route, Switch, useHistory } from "react-router-dom";
+import { useDispatch, useSelector } from "react-redux";
+import { useToasts } from "react-toast-notifications";
 import styled from "styled-components";
 
+import AuthAPI from "api/auth";
+import { succeedTokenCheck } from "store/modules/auth";
 import {
   getValueOnLocalStorage,
   removeValueOnLocalStorage,
+  setValueOnLocalStorage,
 } from "utils/localStorageUtils";
 
 import LoginPage from "pages/AuthPages/LoginPage";
@@ -22,7 +27,104 @@ const Main = styled.main`
 `;
 
 const App = () => {
+  const { addToast } = useToasts();
+  const { isCheckedToken } = useSelector((state) => state.authReducer);
+  const dispatch = useDispatch();
+  const history = useHistory();
+
+  /**
+   * refresh_token으로 refresh를 시도합니다.
+   * 성공할 경우 로그인 처리,
+   * 실패할 경우 로그인 페이지로 이동시키며
+   * local storage에 있는 token을 지웁니다.
+   * 참고 : https://app.diagrams.net/#G1fdPJc3IfiFc6l8OMxSJ2rOJisJsp0k8i
+   * @param {string} token
+   */
+  const refreshAccessToken = async (token) => {
+    try {
+      const res = await AuthAPI.refreshToken(token.refresh_token);
+
+      if (res.status === 200) {
+        setValueOnLocalStorage("hangangToken", res.data);
+        dispatch(succeedTokenCheck({ isLoggedIn: true, token: token }));
+      }
+    } catch (err) {
+      if (err.response.data.code === 6) {
+        removeValueOnLocalStorage("hangangToken");
+        history.push("/login");
+        addToast("토큰이 유효하지 않습니다. 다시 로그인 해주세요.", {
+          appearance: "error",
+          autoDismiss: true,
+        });
+      }
+
+      if (err.response.data.code === 9) {
+        removeValueOnLocalStorage("hangangToken");
+        history.push("/login");
+        addToast(
+          "2주 이내로 로그인 하지 않아 토큰이 만료되었습니다. 다시 로그인 해주세요.",
+          {
+            appearance: "error",
+            autoDismiss: true,
+          }
+        );
+      }
+    }
+  };
+
+  /**
+   * access_token이 유효한지 검사합니다.
+   * 유효하다면 로그인 처리, 유효하지 않다면 에러 코드에 따라
+   * 재로그인, refresh_token으로 token 갱신을 시도합니다.
+   * @param {string} token
+   */
+  const checkValidAccessToken = async (token) => {
+    try {
+      const res = await AuthAPI.authTest(token.access_token);
+      if (res.status === 200) {
+        dispatch(succeedTokenCheck({ isLoggedIn: true, token: token }));
+      }
+    } catch (err) {
+      if (err.response.data.code === 5) {
+        removeValueOnLocalStorage("hangangToken");
+        history.push("/login");
+        addToast("토큰이 유효하지 않습니다. 다시 로그인 해주세요.", {
+          appearance: "error",
+          autoDismiss: true,
+        });
+      }
+
+      if (err.response.data.code === 8) {
+        refreshAccessToken(token);
+      }
+    }
+  };
+
+  /**
+   * LS 가져온 토큰을 검사합니다.
+   * 없을 경우 비로그인처리, 있을 경우 유효한지 검사합니다.
+   */
+  const checkTokenOnLocalStorage = () => {
+    let token = getValueOnLocalStorage("hangangToken");
+    if (!token) {
+      dispatch(succeedTokenCheck({ isLoggedIn: false, token: "" }));
+    } else {
+      checkValidAccessToken(token);
+    }
+  };
+
+  /**
+   * IndexPage가 마운트 될 경우,
+   * 유저의 local storage에서 토큰을 체크합니다.
+   *
+   * 유저가 브라우저를 종료시킨다면 autoLogin 값을 체크하여
+   * false라면 LocalStorage의 token을 말소시킵니다.
+   */
   useEffect(() => {
+    if (!isCheckedToken) {
+      checkTokenOnLocalStorage();
+    }
+
     window.addEventListener("beforeunload", () => {
       const autoLoginKey = getValueOnLocalStorage("didHangangAutoLogin");
       if (!autoLoginKey) {
@@ -41,18 +143,16 @@ const App = () => {
 
   return (
     <Main role="main">
-      <Router>
-        <NavigationContainer />
-        <Switch>
-          <Route path="/" exact component={IndexPage} />
-          <Route path="/login" component={LoginPage} />
-          <Route path="/findpwauth" component={FindPwAuthPage} />
-          <Route path="/findpw" component={FindPwPage} />
-          <Route path="/signupauth" component={SignUpAuthPage} />
-          <Route path="/signup" component={SignUpPage} />
-        </Switch>
-        <FooterContainer />
-      </Router>
+      <NavigationContainer />
+      <Switch>
+        <Route path="/" exact component={IndexPage} />
+        <Route path="/login" component={LoginPage} />
+        <Route path="/findpwauth" component={FindPwAuthPage} />
+        <Route path="/findpw" component={FindPwPage} />
+        <Route path="/signupauth" component={SignUpAuthPage} />
+        <Route path="/signup" component={SignUpPage} />
+      </Switch>
+      <FooterContainer />
     </Main>
   );
 };
